+++ R/as-tsibble.R
@@ -220,6 +220,23 @@ key.tbl_ts <- function(x) {
   attr(x, "key")
 }
 
+#' @rdname key
+#' @export
+unkey <- function(x) {
+  UseMethod("unkey")
+}
+
+#' @rdname key
+#' @export
+unkey.tbl_ts <- function(x) {
+  if (n_keys(x) > 1) {
+    abort("`unkey()` must not be applied to a `tbl_ts` of more than 1 key size.")
+  }
+  attr(x, "key") <- structure(id(), class = "key")
+  attr(x, "key_indices") <- NULL
+  x
+}
+
 `key<-` <- function(x, value) {
   attr(x, "key") <- value
   x